Crypto-TS-Model-master/requirements.txt
# Core requirements
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
tqdm>=4.62.0
pyyaml>=5.4.1

# Technical analysis
ta>=0.10.0

# Optional for GPU support
# cudatoolkit=11.3  # Uncomment if using CUDA

# Development tools
black>=22.3.0  # Code formatting
flake8>=4.0.0  # Linting
pytest>=7.0.0  # Testing
